requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
dnacentersdk==2.0.2
fastjsonschema==2.15.3
Flask==1.1.2
frozenlist==1.3.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
meraki==1.4.3
multidict==6.0.2
PyJWT==1.7.1
python-dotenv==0.14.0
requests==2.24.0
requests-toolbelt==0.9.1
urllib3==1.25.10
webexteamssdk==1.6
Werkzeug==1.0.1
yarl==1.7.2
